backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
supabase==2.4.3
redis[hiredis]==5.0.4
celery==5.4.0
flower==2.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic-settings==2.2.1
python-dotenv==1.0.1
slowapi==0.1.9
fastapi-mail==1.4.1
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0